requirements.txt
gower==0.1.2
numpy==1.24.3
pandas==2.0.2
scikit_learn==1.3.0
scipy
google-vizier[jax]
git+https://github.com/AlvaroGI/optimizing-cd-protocols.git@package
